+++ examples/kraken-interpol/models/index.js
@@ -1,9 +1,0 @@
-'use strict';
-
-
-module.exports = function IndexModel() {
-    return {
-        lang: 'US_en', // Only for example
-        name: 'kraken-interpol'
-    };
-};